requirements.txt
django
djangorestframework
django-passwords
django-model-utils
django-debug-toolbar
